cleanup parser files and split them
